Move personal_cl_functions to cl/functions
