--- include/visiongl/cl/personal_cl_functions.h
@@ -1,5 +1,0 @@
-#include <visiongl/image.hpp>
-
-void vglCl3dMergeZByMean(VglImage* img_input, VglImage* img_output, int number_of_merges);
-void vglCl3dMergeZByMax(VglImage* img_input, VglImage* img_output, int number_of_merges);
-float vglCl9root(float number);
